Multiple changes, see description.
- Fix conflicts between Prayer of Azariah and Song of the
Three Young Men.

These books are (typically) the same, and once upon a
time we could use 'Prayer of Azariah' to refer to them all.
However, we cannot do this anymore, and have to keep
them distinct.

- Fix conflicts between Greek Esther and Additions to Esther.

Same deal as above.

- Added tests for potential cases of the aforementioned.
- Added warning logs for when name fetching finds
these varying names, including the 'default' that we use in code,
as well as warnings for when we encounter a data name that
we don't have a matching book for.