feat: version aliases
This implements a more proper structure to version aliases.

As an example, the NRSV points to the NRSVA because our source has deprecated the NRSV in favor of the NRSVue, but the NRSVA remains the older text.

For the longest while, we had hardcoded checks to shim the NRSVA in whenever it came time to pull the content. This replaces that with a version-agnostic arrangement that handles version aliases across the backend.

In other words:
- Added AliasOf to the Version model.
- Redirect to the aliased version in reference generation where applicable.
- Use the aliased version when calling /booklist.
- Added an indicator to /versioninfo when version is an alias.