feat!: user-installed app support, implement /verse for user-installed app
Adds support for user-installed apps, with the new `/verse` command for verse referencing in servers without BibleBot. This essentially does the same behavior as making a message with a verse reference in a server with BibleBot, except a message will return if nothing is returned.

Server-specific commands are now disabled for those using it in a user app context. Additionally, most command responses are only visible for the user running the command in a user app context.

While I was here, I tidied up much of the procedures for talking to the backend and sending out responses. There's still a lot of work to do in terms of making the linter(s) stop freaking out, but at least some of it is cleaner.

Fixes #416 and #418.